+++ nodes/conversion/src/lib.rs
@@ -1,14 +1,14 @@
 use mizer_node_api::*;
 
 pub struct ConvertToDmxNode {
-    channel: u8,
+    channel: u16,
     universe: u16,
     outputs: Vec<DmxSender>,
     inputs: Vec<NumericChannel>,
 }
 
 impl ConvertToDmxNode {
-    pub fn new(universe: Option<u16>, channel: Option<u8>) -> Self {
+    pub fn new(universe: Option<u16>, channel: Option<u16>) -> Self {
         let universe = universe.unwrap_or_default();
         let channel = channel.unwrap_or_default();
         log::trace!("New ConvertToDmxNode({}.{:0>3})", universe + 1, channel + 1);
